backend/intelligence-service/requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Pydantic for data validation
pydantic==2.5.0

# Anthropic Claude API
anthropic==0.39.0
httpx==0.25.2

# Retry logic
tenacity==8.2.3

# Environment variables
python-dotenv==1.0.0

# String similarity (for matching)
python-Levenshtein==0.25.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# CORS is already included in FastAPI - no need for separate package